warehouse_project_server/requirements.txt
asgiref==3.7.2
Django==3.2.5
django-cors-headers==4.2.0
django-cron==0.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
